refactor: extract EventBusMetrics into separate file
